subjects/pyproject.toml
[tool.poetry]
name = "subjects"
version = "0.1.0"
description = ""
authors = ["Clara Baumgarten <clara.odebrecht@olist.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.1.6"
djangorestframework = "^3.12.2"
prettyconf = "^2.2.1"

[tool.poetry.dev-dependencies]
pytest = "^4.6"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
